fix(runtime): input() returned 0 on bad input, hung at EOF, truncated to 32 bits (#308)
PLAN_v1.21 §2 logged one defect here: `input()` type-checks clean and prints 0
for a non-numeric line - a silent wrong answer at exit 0, on the first line of
any tutorial program. Probing it found five more, all from ONE root cause: a
token reader (`scanf("%d")`) with the wrong C return types, which could not
validate or bound what it consumed.

  (a) a non-numeric line returned 0                 ("hello" -> 0)
  (b) the drain loop `while (getchar() != '\n')` never terminated at EOF,
      because getchar() keeps returning EOF -- an UNBOUNDED HANG on empty stdin
  (c) a trailing tail was swallowed                 ("12abc" -> 12)
  (d) the newline was left in the buffer, so a FOLLOWING input_line() returned
      "" instead of the next line
  (e) the C return type was `int` while a Wyn int is 64-bit (4294967297 -> 1),
      and a 64-bit overflow gave -1 rather than an error
  (f) the C return type was `float` while a Wyn float is a C double
      (0.1234567890123 -> 0.12345679104328156) -- the same class as the
      v1.20.0 float-printing round-trip fix, surviving in the stdin path

Every one was a wrong answer at exit 0 or a hang; (b) is the worst, since a
hang in a pipeline is unbounded.

Fix: read one line via input_line() -- already RC-correct, growable and
CRLF-safe -- then validate the WHOLE line, exactly as str_parse_int (`to_int`)
does, and reuse its panic posture: report the offending value on stderr, exit 1,
and let WYN_LENIENT=1 restore the old return-0 behaviour. Return long long /
double to match Wyn's int and float. This makes input() mean int(input_line()),
which is why (d) disappears rather than needing its own fix.

No signature change at the Wyn level (`input()` is still int-typed), so this is
not a semver break; only the wrong C types underneath it change.

Verification
- 13 new assertions in tests/errors/run_silent_wrong_test.sh (the existing
  silent-wrong-answer batch, which already owned input_line's section): 29/0.
  Every case is wrapped in alarm() because (b) was a hang -- a regression must
  FAIL, not wedge the suite.
- MUTATION-VERIFIED: reverting only the runtime files fails all 9 new
  behavioural assertions while the 4 guards (valid int, valid float,
  WYN_LENIENT, and a USER fn named `input` shadowing the builtin) still pass,
  so the tests are specific rather than vacuous.
- make test EXIT=0: BDD 289/0, golden-C 30/0, stdlib 68/68 0 hang.
- examples 64/64 `wyn check` AND 64/64 `wyn build`.
- ASan runtime lib, happy path plus all six error branches under WYN_LENIENT=1
  (which continues past the wyn_rc_release, the risky path): no reports.
- scripts/flagship_apps_gate.sh full run: 8/8 PASS at recorded baselines.

Note for reviewers: nothing in 1052 corpus .wyn files calls input(), and no doc
mentions it -- the same zero-usage signal as #301/#304, and why six defects sat
in the most basic stdin path undetected.

+++ src/wyn_runtime_slim.h
@@ -560,8 +560,11 @@ void print_value(WynValue v);
 void print_hex(int x);
 void print_bin(int x);
 void print_debug(const char* label, int val);
-int input();
-float input_float();
+// long long / double, not int / float: a Wyn int is 64-bit and a Wyn float is a
+// C double, so the narrower C types silently truncated (4294967297 -> 1) and lost
+// precision. Must match the definitions in wyn_runtime.h.
+long long input();
+double input_float();
 char* input_line();
 void printf_wyn(const char* format, ...);
 char* string_format(const char* format, ...);
